commune/modules/eth/requirements.txt
web3
eth-account
eth-utils
eth-keys
pycryptodome
py-solc-x